feat: working version of line

+++ src/pixels/comet.cpp
@@ -27,7 +27,7 @@ void updateComet(Adafruit_NeoPixel& pixels, uint16_t state[]) {
     }
 
     for (int16_t row = getParameter(PARAM_NB_ROWS) - 1; row >= 0; row--) {
-      for (uint8_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
+      for (uint16_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
            column++) {
         uint16_t led = getLedIndex(row, column);
         if ((state[led] == 2) ||
@@ -62,7 +62,7 @@ void updateComet(Adafruit_NeoPixel& pixels, uint16_t state[]) {
   }
 
   uint8_t row = random(0, getParameter(PARAM_NB_ROWS));
-  uint8_t column = random(0, getParameter(PARAM_NB_COLUMNS));
+  uint16_t column = random(0, getParameter(PARAM_NB_COLUMNS));
   int16_t led;
   uint8_t direction = getDirection();
   switch (direction) {
