+++ src/pixels/wave.cpp
@@ -29,7 +29,7 @@ void updateWave(Adafruit_NeoPixel& pixels) {
     case 2:
       for (uint8_t row = 0; row < getParameter(PARAM_NB_ROWS); row++) {
         uint16_t led = 0;
-        for (uint8_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
+        for (uint16_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
              column++) {
           led = getLedIndex(row, column);
           int16_t nextLedIndex = getNextLedIndex(row, column, direction);
@@ -41,7 +41,7 @@ void updateWave(Adafruit_NeoPixel& pixels) {
       }
       break;
     case 3:
-      for (uint8_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
+      for (uint16_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
            column++) {
         uint16_t led = 0;
         for (int16_t row = getParameter(PARAM_NB_ROWS) - 1; row >= 0; row--) {
@@ -55,7 +55,7 @@ void updateWave(Adafruit_NeoPixel& pixels) {
       }
       break;
     case 4:
-      for (uint8_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
+      for (uint16_t column = 0; column < getParameter(PARAM_NB_COLUMNS);
            column++) {
         uint16_t led = 0;
         for (uint8_t row = 0; row < getParameter(PARAM_NB_ROWS); row++) {
